pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "new-songs-enjoyer"
version = "0.1.0"
description = "Crafts you a playlist to discover new songs based on preferences using Last.fm API"
readme = "README.md"
license = "BUSL-1.1"
requires-python = ">=3.8"
authors = [
    {name = "ArcadiaWorks"}
]
keywords = ["music", "lastfm", "playlist", "recommendations"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
]

dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/ArcadiaWorks/new_songs_enjoyer"
Repository = "https://github.com/ArcadiaWorks/new_songs_enjoyer.git"

[project.scripts]
new-songs-enjoyer = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = ["main.py", "config.py", "entities/", "templates/", "adapter/"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.5",
]
